refactor: extract ColladaParser, ColladaJointExtractor, ColladaMeshProcessor from JointedModelColladaExporter
Reduce JointedModelColladaExporter from 1181 to 414 lines by extracting:
- ColladaJointExtractor (76 lines): skeleton tree traversal and joint nodes
- ColladaParser (174 lines): COLLADA XML assembly, assets, effects, textures
- ColladaMeshProcessor (457 lines): geometry, skins, controllers, scene nodes

All 171 tests pass including 49 TDD decomposition tests.
Public API preserved — no caller changes needed.
